+++ satsim/satsim.py
@@ -29,7 +29,13 @@ from satsim.geometry.sstr7 import query_by_los
 from satsim.geometry.csvsc import query_by_los as csvsc_query_by_los
 from satsim.geometry.sgp4 import create_sgp4
 from satsim.geometry.ephemeris import create_ephemeris_object
-from satsim.geometry.astrometric import create_topocentric, gen_track, get_los, get_los_azel
+from satsim.geometry.astrometric import (
+    create_topocentric,
+    gen_track,
+    get_los,
+    get_los_azel,
+    get_analytical_los,
+)
 from satsim.geometry.greatcircle import GreatCircle
 from skyfield.toposlib import _ltude
 from satsim.geometry.photometric import model_to_mv
@@ -456,6 +462,9 @@ def image_generator(ssp, output_dir='.', output_debug=False, dir_debug='./Debug'
 
     if 'analytical_obs' not in ssp['sim']:
         ssp['sim']['analytical_obs'] = False
+
+    if 'analytical_obs_frame' not in ssp['sim']:
+        ssp['sim']['analytical_obs_frame'] = 'geocentric'
 
     if 'psf_sample_frequency' not in ssp['sim']:
         ssp['sim']['psf_sample_frequency'] = 'once'
@@ -1122,6 +1131,9 @@ def _gen_objects(ssp, render_mode,
                     fliplr=ssp['fpa']['flip_left_right'],
                     az=az,
                     el=el,
+                    deflection=ssp['sim']['enable_deflection'],
+                    aberration=ssp['sim']['enable_light_transit'],
+                    stellar_aberration=ssp['sim']['enable_stellar_aberration'],
                 )
             except Exception:
                 logger.exception("Error propagating target. {}".format(o))
@@ -1184,13 +1196,11 @@ def _gen_objects(ssp, render_mode,
             oppp = np.concatenate((oppp, opp))
 
         if obs_cache[i] is not None:
-            ra_mid, dec_mid, _, _, _, _ = get_los(
+            ra_mid, dec_mid, _ = get_analytical_los(
                 observer,
                 target,
                 ts_mid,
-                deflection=ssp['sim']['enable_deflection'],
-                aberration=ssp['sim']['enable_light_transit'],
-                stellar_aberration=False,
+                frame=ssp['sim']['analytical_obs_frame']
             )
             ra_true, dec_true, _, _, _, _ = get_los(
                 observer,
@@ -1225,6 +1235,7 @@ def _gen_objects(ssp, render_mode,
         if ra_mid is not None and dec_mid is not None:
             entry['ra_obs'] = ra_mid
             entry['dec_obs'] = dec_mid
+            entry['obs_frame'] = ssp['sim']['analytical_obs_frame']
             entry['ra'] = ra_true
             entry['dec'] = dec_true
 
@@ -1273,7 +1284,28 @@ def _calculate_star_position_and_motion(ssp, astrometrics,
     else:
         logger.error('Unknown track mode: {}.'.format(track_mode))
 
-    [rr0, rr1], [cc0, cc1], drr, dcc, _ = gen_track(h_fpa_pad_os, w_fpa_pad_os, y_fov_pad, x_fov_pad, observer, track, track_target, [0], ts_collect_start, [t_start, t_end], star_rot, 1, track_mode, flipud=ssp['fpa']['flip_up_down'], fliplr=ssp['fpa']['flip_left_right'], az=az, el=el)
+    [rr0, rr1], [cc0, cc1], drr, dcc, _ = gen_track(
+        h_fpa_pad_os,
+        w_fpa_pad_os,
+        y_fov_pad,
+        x_fov_pad,
+        observer,
+        track,
+        track_target,
+        [0],
+        ts_collect_start,
+        [t_start, t_end],
+        star_rot,
+        1,
+        track_mode,
+        flipud=ssp['fpa']['flip_up_down'],
+        fliplr=ssp['fpa']['flip_left_right'],
+        az=az,
+        el=el,
+        deflection=enable_deflection,
+        aberration=enable_light_transit,
+        stellar_aberration=enable_stellar_aberration,
+    )
     star_tran_os = [-drr, -dcc]  # stars move in the opposite direction of target
     star_rot_rate = 0  # TODO